requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
docopt==0.6.2
idna==3.3
pipreqs==0.4.11
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
yarg==0.1.9

selenium~=3.141.0
bs4~=0.0.1